Set src and dst port to 0 for non-TCP/UDP packets
The source and destination ports are used to find the right bucket in the hash.
Currently they are only set when a TCP or UDP packet was detected and otherwise
the old (from previous packet) value is used. This is problematic when in the
future non-UDP/TCP traffic should be dumped because it will end up in different
pcap files.