Handle absolute paths in as output directory
Always assumed that the first directory in the output directory parameter is
under the cwd when doing the output directories check/creation. But it didn't
use the same assumption when doing the file opening.

When the user specified as output dir /foo and the cwd was /bar then pkt2flow
checked whether /bar/foo exists and created it when necessary. It then tried
to open the files under /foo/ (which fails when it wasn't created earlier).

This behavior was changed to use the cwd when the output path doesn't start
with a "/" and omitting the cwd when it starts with a "/"