reclassify flow status after FLOW_TIMEOUT
A flow which is marked as timeout and restarted again gets the status
STS_UNSET. This results in the output path "others/". This can often happen
with UDP packets like DHCP because they aren't send very frequenty. The
"others"/STS_UNSET classification is wrong in this case because it is known
that this flow is UDP. This is especially important when the "others" output
type is not enabled.

The flow should be reclassified instead to calculate the correct output folder.